src/main/java/com/kamil/BookingSystem/Repositories/ClientRepository.java
package com.kamil.BookingSystem.Repositories;

import com.kamil.BookingSystem.Entity.Client;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
import org.springframework.data.jpa.repository.Query;
import org.springframework.stereotype.Repository;

import java.util.Optional;

@Repository
public interface ClientRepository extends JpaRepository<Client,Long> {
    @Query("SELECT c FROM Client c WHERE c.email = ?1")
    public Optional<Client> getByEmail(String email);

}
